setup + documentation + extra spaces

+++ build_setup_help_on_windows.bat
@@ -1,59 +1,77 @@
-rem we remove everything from dist
 echo off
-del /Q dist\*.*
+IF EXIST dist del /Q dist\*.*
 
-rem unittests with python 3.4
+IF NOT EXIST c:\Python34 GOTO checkinstall64:
 
-IF EXIST c:\Python34vir GOTO next:
+:checkinstall:
+IF EXIST c:\Python34vir GOTO nexta:
 mkdir c:\Python34vir
 
-:next:
-IF EXIST c:\Python34vir\install GOTO fullsetup:
+:nexta:
+IF EXIST c:\Python34vir\install GOTO fullsetupa:
 c:\Python34\Scripts\virtualenv c:\Python34vir\install --system-site-packages
 if %errorlevel% neq 0 exit /b %errorlevel%
 
-:fullsetup:
-
+:fullsetupa:
 echo #######################################################
 c:\Python34vir\install\Scripts\python -u setup.py install
 if %errorlevel% neq 0 exit /b %errorlevel%
 echo #######################################################
 
-set pythonexe="c:\Python34\python"
+
+:checkinstall64:
+IF EXIST c:\Python34_64vir GOTO nextb:
+mkdir c:\Python34_64vir
+
+:nextb:
+IF EXIST c:\Python34_64vir\install GOTO fullsetupb:
+c:\Python34_x64\Scripts\virtualenv c:\Python34_64vir\install --system-site-packages
+if %errorlevel% neq 0 exit /b %errorlevel%
+
+:fullsetupb:
+echo #######################################################
+c:\Python34_64vir\install\Scripts\python -u setup.py install
+if %errorlevel% neq 0 exit /b %errorlevel%
+echo #######################################################
+
+:utpy34_64:
+set pythonexe="c:\Python34_x64\python"
 %pythonexe% -u setup.py clean_space
 if %errorlevel% neq 0 exit /b %errorlevel%
 %pythonexe% -u setup.py unittests
 if %errorlevel% neq 0 exit /b %errorlevel%
 echo #######################################################
 
-rem python 3.3
-
-set pythonexe="c:\python33_x64\python"
-%pythonexe% clean_pyd.py
+:setup33_64:
+IF NOT EXIST c:\Python33_x64 GOTO setup33:
+set pythonexe="c:\Python33_x64\python"
+%pythonexe% setup.py clean_pyd
 %pythonexe% setup.py build bdist_wininst --plat-name=win-amd64
 if %errorlevel% neq 0 exit /b %errorlevel%
 echo #######################################################
 
-set pythonexe="c:\python33\python"
-%pythonexe% clean_pyd.py
+:setup33:
+IF NOT EXIST c:\Python33 GOTO setup34_64:
+set pythonexe="c:\Python33\python"
+%pythonexe% setup.py clean_pyd
 %pythonexe% setup.py sdist --formats=gztar,zip --verbose
 if %errorlevel% neq 0 exit /b %errorlevel%
 %pythonexe% setup.py bdist_wininst
 if %errorlevel% neq 0 exit /b %errorlevel%
 echo #######################################################
 
-rem python 3.4
-
-set pythonexe="c:\python34_x64\python"
-%pythonexe% clean_pyd.py
+:setup34_64:
+set pythonexe="c:\Python34_x64\python"
+%pythonexe% setup.py clean_pyd
 %pythonexe% setup.py build bdist_wininst --plat-name=win-amd64
 if %errorlevel% neq 0 exit /b %errorlevel%
 %pythonexe% setup.py build bdist_msi --plat-name=win-amd64
 if %errorlevel% neq 0 exit /b %errorlevel%
 echo #######################################################
 
-set pythonexe="c:\python34\python"
-%pythonexe% clean_pyd.py
+:setup_msi:
+set pythonexe="c:\Python34_x64\python"
+%pythonexe% setup.py clean_pyd
 %pythonexe% setup.py sdist --formats=gztar,zip --verbose
 if %errorlevel% neq 0 exit /b %errorlevel%
 %pythonexe% setup.py bdist_wininst
@@ -62,8 +80,12 @@ if %errorlevel% neq 0 exit /b %errorlevel%
 if %errorlevel% neq 0 exit /b %errorlevel%
 echo #######################################################
 
-rem help
+:documentation:
+%pythonexe% -u setup.py build_sphinx
+if %errorlevel% neq 0 exit /b %errorlevel%
+echo #######################################################
 
+:presentation:
 %pythonexe% -u setup.py build_pres
 if %errorlevel% neq 0 exit /b %errorlevel%
 %pythonexe% -u setup.py build_pres_2A
@@ -74,6 +96,7 @@ if %errorlevel% neq 0 exit /b %errorlevel%
 if %errorlevel% neq 0 exit /b %errorlevel%
 echo #######################################################
 
+:copyfiles:
 if not exist dist\html mkdir dist\html
 if not exist dist\html2 mkdir dist\html2
 if not exist dist\html3 mkdir dist\html3
